+++ lib/input/constructor.go
@@ -65,6 +65,7 @@ const (
 	TypeAMQP            = "amqp"
 	TypeAMQP09          = "amqp_0_9"
 	TypeAMQP1           = "amqp_1"
+	TypeBloblang        = "bloblang"
 	TypeBroker          = "broker"
 	TypeDynamic         = "dynamic"
 	TypeFile            = "file"
@@ -108,6 +109,7 @@ type Config struct {
 	AMQP            reader.AMQPConfig            `json:"amqp" yaml:"amqp"`
 	AMQP09          reader.AMQP09Config          `json:"amqp_0_9" yaml:"amqp_0_9"`
 	AMQP1           reader.AMQP1Config           `json:"amqp_1" yaml:"amqp_1"`
+	Bloblang        BloblangConfig               `json:"bloblang" yaml:"bloblang"`
 	Broker          BrokerConfig                 `json:"broker" yaml:"broker"`
 	Dynamic         DynamicConfig                `json:"dynamic" yaml:"dynamic"`
 	File            FileConfig                   `json:"file" yaml:"file"`
@@ -152,6 +154,7 @@ func NewConfig() Config {
 		AMQP:            reader.NewAMQPConfig(),
 		AMQP09:          reader.NewAMQP09Config(),
 		AMQP1:           reader.NewAMQP1Config(),
+		Bloblang:        NewBloblangConfig(),
 		Broker:          NewBrokerConfig(),
 		Dynamic:         NewDynamicConfig(),
 		File:            NewFileConfig(),